Blog v2.0
This version has flask-login library implemented in.

Now you need to signup first to be able to create a post.

Functions signup, signin and logout were added.

This a v2.0 of the project. Keeping things simple, this a good template
to a start a big project.

Visit us at @codenoschool (github, youtube, twitter).